# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Simple API Restaurant",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
        "/allergens": {
            "get": {
                "tags": [
                    "allergens"
                ],
                "summary": "Lists all allergens with their associated meals",
                "description": "Return a list of Allergens",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Please provide correct data"
                    }
                }
            },
            "post": {
                "tags": [
                    "allergens"
                ],
                "summary": "Create one Allergen",
                "description": "Create a allergen resource",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Title of Allergen",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Entity, validation fails."
                    }
                }
            }
        },
        "/allergens/{allergen}": {
            "get": {
                "tags": [
                    "allergens"
                ],
                "summary": "Read one Allergen",
                "description": "Return a Allergen resource",
                "parameters": [
                    {
                        "name": "allergen",
                        "in": "path",
                        "description": "Id of allergen",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Please provide correct data"
                    }
                }
            },
            "put": {
                "tags": [
                    "allergens"
                ],
                "summary": "Update one Allergen",
                "description": "Update an Allergen and return name.",
                "parameters": [
                    {
                        "name": "allergen",
                        "in": "path",
                        "description": "Id of allergen",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Title of allergen",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity, validation fails."
                    }
                }
            },
            "delete": {
                "tags": [
                    "allergens"
                ],
                "summary": "Delete one Allergen.",
                "description": "A allergen resource will be deleted",
                "parameters": [
                    {
                        "name": "allergen",
                        "in": "path",
                        "description": "Id of allergen",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Resource deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/ingredients": {
            "get": {
                "tags": [
                    "ingredients"
                ],
                "summary": "Lists all allergens with their associated meals",
                "description": "Return a list of Ingredients",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Please provide correct data"
                    }
                }
            },
            "post": {
                "tags": [
                    "ingredients"
                ],
                "summary": "Create one Ingredient",
                "description": "Return a Ingredients name and description",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Title of ingredient",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Entity, validation fails."
                    }
                }
            }
        },
        "/ingredients/{ingredient}": {
            "get": {
                "tags": [
                    "ingredients"
                ],
                "summary": "Read one Ingredient",
                "description": "Return a Ingredient resource",
                "parameters": [
                    {
                        "name": "ingredient",
                        "in": "path",
                        "description": "Id of ingredient",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Please provide correct data"
                    }
                }
            },
            "put": {
                "tags": [
                    "ingredients"
                ],
                "summary": "Update one Ingredent",
                "description": "Return a Ingredent name and description",
                "parameters": [
                    {
                        "name": "ingredient",
                        "in": "path",
                        "description": "Id of ingredient",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of ingredient",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity, validation fails."
                    }
                }
            },
            "delete": {
                "tags": [
                    "ingredients"
                ],
                "summary": "Delete one Ingredent.",
                "description": "A ingredient resource will be deleted",
                "parameters": [
                    {
                        "name": "ingredient",
                        "in": "path",
                        "description": "Id of ingredient",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Resource deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
...
}